* TimeMonColors.m           float -> CGFloat         * Percentages.m           Add alpha to the default colors to actually show something
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3059 d25b4930-3292-e311-b074-0010c6de5da8

+++ ported-apps/Util/TimeMon/TimeMonColors.m
@@ -66,7 +66,7 @@ static const id titles[] = {
 	else 
 	  {
             NSString *key = [NSString stringWithFormat:@"%@Gray",titles[i]];
-            float color = [defaults floatForKey:key];
+            CGFloat color = [defaults floatForKey:key];
             [tCell setColor:[NSColor colorWithCalibratedWhite:color alpha:1.0]];
 	  }
         [tCell setBezeled:YES];
